Started the power-ups code. See description for more details\/\/
created powerups class and put in code to the enemy ship 2 class that
will be dropping the powerup  since the others either will move in
seperate patterns or are too far away. all code for movement and hit
detection is complete i just need to code the methods that will actually
give the player his attributes and give them images.